Moved fingerprinting function within hdf-converters
