+++ apps/frontend/src/store/report_intake.ts
@@ -2,10 +2,10 @@
  * Reads and parses inspec files
  */
 
-import {InspecDataModule} from '@/store/data_store';
+import { InspecDataModule } from '@/store/data_store';
 import Store from '@/store/store';
-import {Tag} from '@/types/models';
-import {read_file_async} from '@/utilities/async_util';
+import { Tag } from '@/types/models';
+import { read_file_async } from '@/utilities/async_util';
 import {
   ASFFResults as ASFFResultsMapper,
   BurpSuiteMapper,
@@ -33,11 +33,12 @@ import {
   convertFile,
   ExecJSON
 } from 'inspecjs';
+import { guessType } from '../../../../libs/hdf-converters/src/utils/check-input-format' // TODO - figure out how we want to import this
 import _ from 'lodash';
-import {v4 as uuid} from 'uuid';
-import {Action, getModule, Module, VuexModule} from 'vuex-module-decorators';
-import {FilteredDataModule} from './data_filters';
-import {SnackbarModule} from './snackbar';
+import { v4 as uuid } from 'uuid';
+import { Action, getModule, Module, VuexModule } from 'vuex-module-decorators';
+import { FilteredDataModule } from './data_filters';
+import { SnackbarModule } from './snackbar';
 
 /** Each FileID corresponds to a unique File in this store */
 export type FileID = string;
@@ -112,30 +113,6 @@ export type ExecJSONLoadOptions = {
   data: ExecJSON.Execution;
 };
 
-// Fields to look for inside of JSON structures to determine type before passing to hdf-converters
-export const fileTypeFingerprints = {
-  asff: ['Findings', 'AwsAccountId', 'ProductArn'],
-  fortify: ['FVDL', 'FVDL.EngineData.EngineVersion', 'FVDL.UUID'],
-  ionchannel: ['analysis_id', 'team_id', 'source', 'trigger_hash'],
-  jfrog: ['total_count', 'data'],
-  nikto: ['banner', 'host', 'ip', 'port', 'vulnerabilities'],
-  sarif: ['$schema', 'version', 'runs'],
-  snyk: [
-    'projectName',
-    'policy',
-    'summary',
-    'vulnerabilities',
-    'vulnerabilities[0].identifiers'
-  ],
-  twistlock: [
-    'results[0].complianceDistribution',
-    'results[0].vulnerabilityDistribution',
-    'results[0].collections',
-    'results[0].digest'
-  ],
-  zap: ['@generated', '@version', 'site']
-};
-
 @Module({
   namespaced: true,
   dynamic: true,
@@ -182,9 +159,9 @@ export class InspecIntake extends VuexModule {
               filename: `${filename
                 .replace(/.json/gi, '')
                 .replace(/.nessus/gi, '')}-${_.get(
-                evaluation,
-                'platform.target_id'
-              )}.${originalFileType}`
+                  evaluation,
+                  'platform.target_id'
+                )}.${originalFileType}`
             });
           })
         );
@@ -239,7 +216,7 @@ export class InspecIntake extends VuexModule {
       convertOptions.fileOptions.filename ||
       'No Filename';
     // If the data passed is valid json, try to match up known keys
-    const typeGuess = await this.guessType({
+    const typeGuess = await guessType({
       data: convertOptions.data,
       filename: filename
     });
@@ -284,72 +261,6 @@ export class InspecIntake extends VuexModule {
   }
 
   @Action
-  async guessType(guessOptions: {
-    data: string;
-    filename: string;
-  }): Promise<string> {
-    try {
-      const parsed = JSON.parse(guessOptions.data);
-      const object = Array.isArray(parsed) ? parsed[0] : parsed;
-      // Find the fingerprints that have the most matches
-      const fingerprinted = Object.entries(fileTypeFingerprints).reduce(
-        (a, b) => {
-          return a[1].filter((value) => _.get(object, value)).length >
-            b[1].filter((value) => _.get(object, value)).length
-            ? {...a, count: a[1].filter((value) => _.get(object, value)).length}
-            : {
-                ...b,
-                count: b[1].filter((value) => _.get(object, value)).length
-              };
-        }
-      ) as unknown as string[] & {count: number};
-      const result = fingerprinted[0];
-      if (fingerprinted.count !== 0) {
-        return result;
-      }
-    } catch {
-      const splitLines = guessOptions.data.trim().split('\n');
-      // If we don't have valid json, look for known strings inside the file text
-      if (guessOptions.filename.toLowerCase().endsWith('.nessus')) {
-        return 'nessus';
-      } else if (
-        guessOptions.data.match(/xmlns.*http.*\/xccdf/) || // Keys matching (hopefully) all xccdf formats
-        guessOptions.filename.toLowerCase().indexOf('xccdf') !== -1
-      ) {
-        return 'xccdf';
-      } else if (guessOptions.data.match(/<netsparker-.*generated.*>/)) {
-        return 'netsparker';
-      } else if (
-        guessOptions.data.indexOf('"AwsAccountId"') !== -1 &&
-        guessOptions.data.indexOf('"SchemaVersion"') !== -1
-      ) {
-        return 'asff';
-      } else if (guessOptions.data.indexOf('issues burpVersion') !== -1) {
-        return 'burp';
-      } else if (guessOptions.data.indexOf('scoutsuite_results') !== -1) {
-        return 'scoutsuite';
-      } else if (
-        guessOptions.data.indexOf('Policy') !== -1 &&
-        guessOptions.data.indexOf('Job Name') !== -1 &&
-        guessOptions.data.indexOf('Check ID') !== -1 &&
-        guessOptions.data.indexOf('Result Status')
-      ) {
-        return 'dbProtect';
-      } else if (
-        splitLines[0].includes('Hostname') &&
-        splitLines[0].includes('Distro') &&
-        splitLines[0].includes('CVE ID') &&
-        splitLines[0].includes('Compliance ID') &&
-        splitLines[0].includes('Type') &&
-        splitLines[0].includes('Severity')
-      ) {
-        return 'prisma';
-      }
-    }
-    return '';
-  }
-
-  @Action
   async detectAndLoadPredefinedJSON() {
     // On page load, check for the flag to load the preloaded JSON file
     const queryString = window.location.search;
@@ -361,8 +272,8 @@ export class InspecIntake extends VuexModule {
             Accept: 'application/json'
           }
         })
-        .then(async ({data}) => {
-          data.forEach(async (file: {filename: string; data: string}) => {
+        .then(async ({ data }) => {
+          data.forEach(async (file: { filename: string; data: string }) => {
             InspecIntakeModule.loadFile({
               file: new File([new Blob([file.data])], file.filename)
             });
